restructure project to bring tests into one directory
