+++ tests/consequence_prediction/repeat_expansion_variants/resources/deletion.xml.gz
