+++ tests/consequence_prediction/vep_mapping_pipeline/resources/input.xml.gz
