+++ tests/consequence_prediction/vep_mapping_pipeline/resources/output_mappings.tsv
