+++ tests/eva_cttv_pipeline/evidence_string_generation/resources/expected_genetics_evidence_string.json
