+++ tests/eva_cttv_pipeline/evidence_string_generation/resources/ignore_file.txt
