+++ tests/eva_cttv_pipeline/evidence_string_generation/resources/snp2gene_extract.tsv
