+++ tests/eva_cttv_pipeline/evidence_string_generation/resources/string_to_ontology_mappings.tsv
