+++ tests/eva_cttv_pipeline/evidence_string_generation/resources/test_clinvar_record.xml.gz
