+++ tests/eva_cttv_pipeline/evidence_string_generation/test_clinvar.py
@@ -2,7 +2,7 @@ import unittest
 
 
 from eva_cttv_pipeline.evidence_string_generation import consequence_type as CT
-from tests.evidence_string_generation import config
+from tests.eva_cttv_pipeline.evidence_string_generation import config
 
 
 class TestClinvarRecord(unittest.TestCase):
