+++ tests/eva_cttv_pipeline/trait_mapping/resources/test_oxo_data.py
