+++ tests/eva_cttv_pipeline/trait_mapping/test_create_efo_table.py
