+++ tests/eva_cttv_pipeline/trait_mapping/test_trait.py
